requirements.txt
pandas==2.0.3 
numpy==1.23.5
matplotlib==3.6.2
statsmodels==0.14.1
alpaca-py==0.33.1
alpaca-trade-api==3.2.0
requests===2.32.0